Removed route class and them as a Coordinates.
Removed the isRoute() from the interface.

Using instance of for checking the instance of object to print the grid.

Removed person statistic class and added all under person umbrella.

Invariants were added inside of Map class to check if they should be within the Grid.

List of activities was added inside the map and was looped to for routes coorditates to stay within grid.

Invariants were added to the private methods inisde the map to check as per the feedback.

in gears compares to , was given a check too.

What was learnt:

* We can have an inital coordinate for our route which was passed as parametere.

* An interface can be created empty for our own sake.

*We could have removed addCoordinate method from interface.

*YAGNI- IF you don't need it don't out it, like Route class, I added becaues I thought to approach an OOPS approach.