One new interfaces were added which include an interface for all objects which required coordiantes on the map that were Route and Obstacle.
Other interface was changed to "MARKUP INTERFACE " only which marks the stuff on map, Obstacle, Route, EMPTY

New class EMPTY, because a null state as said by franklin is always for something which is not correct, so a new class was introduced as EMPTY which was assigned inside the map constructor and all spaces were assigned empty before we add anything on map.

No coordiates were given in empty as per YAGNI, and no invariants too.